+++ usaspending_api/awards/admin.py
@@ -1,8 +1,8 @@
 from django.contrib import admin
-from usaspending_api.awards.models import Award
+from usaspending_api.awards.models import FinancialAccountsByAwardsTransactionObligations
 
 
-@admin.register(Award)
+@admin.register(FinancialAccountsByAwardsTransactionObligations)
 class AwardAdmin(admin.ModelAdmin):
 
     pass
