+++ usaspending_api/financial_activities/models.py
@@ -1,19 +1,15 @@
 from django.db import models
 from usaspending_api.accounts.models import AppropriationAccountBalances
-from usaspending_api.awards.models import RefProgramActivity
-from usaspending_api.awards.models import RefObjectClassCode
-
-# Create your models here.
-
-# AJ 09/07/2016...Added 5 tables below for financial_activities
+from usaspending_api.references.models import RefProgramActivity
+from usaspending_api.references.models import RefObjectClassCode
 
 class FinancialAccountsByProgramActivityObjectClass(models.Model):
     financial_accounts_by_program_activity_object_class_id = models.AutoField(primary_key=True)
     program_activity_name = models.CharField(max_length=164)
-    program_activity_code = models.ForeignKey('RefProgramActivity', models.DO_NOTHING, db_column='program_activity_code')
-    object_class = models.ForeignKey('RefObjectClassCode', models.DO_NOTHING, db_column='object_class')
+    program_activity_code = models.ForeignKey(RefProgramActivity, models.DO_NOTHING, db_column='program_activity_code')
+    object_class = models.ForeignKey(RefObjectClassCode, models.DO_NOTHING, db_column='object_class')
     by_direct_reimbursable_fun = models.CharField(max_length=1)
-    appropriation_account_balances = models.ForeignKey('AppropriationAccountBalances', models.DO_NOTHING)
+    appropriation_account_balances = models.ForeignKey(AppropriationAccountBalances, models.DO_NOTHING)
     ussgl480100_undelivered_or_fyb = models.DecimalField(max_digits=21, decimal_places=0)
     ussgl480100_undelivered_or_cpe = models.DecimalField(max_digits=21, decimal_places=0)
     ussgl483100_undelivered_or_cpe = models.DecimalField(max_digits=21, decimal_places=0)
