+++ usaspending_api/references/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ReferencesConfig(AppConfig):
+    name = 'references'
